Moved everything into sql.py
