+++ python/pyspark/rdd.py
@@ -1387,95 +1387,6 @@ class PipelinedRDD(RDD):
     def _is_pipelinable(self):
         return not (self.is_cached or self.is_checkpointed)
 
-class Row(dict):
-    """
-    An extended L{dict} that takes a L{dict} in its constructor, and exposes those items as fields.
-
-    >>> r = Row({"hello" : "world", "foo" : "bar"})
-    >>> r.hello
-    'world'
-    >>> r.foo
-    'bar'
-    """
-
-    def __init__(self, d):
-        d.update(self.__dict__)
-        self.__dict__ = d
-        dict.__init__(self, d)
-
-class SchemaRDD(RDD):
-    """
-    An RDD of Row objects that has an associated schema. The underlying JVM object is a SchemaRDD,
-    not a PythonRDD, so we can utilize the relational query api exposed by SparkSQL.
-
-    For normal L{RDD} operations (map, count, etc.) the L{SchemaRDD} is not operated on directly, as
-    it's underlying implementation is a RDD composed of Java objects. Instead it is converted to a
-    PythonRDD in the JVM, on which Python operations can be done.
-    """
-
-    def __init__(self, jschema_rdd, sql_ctx):
-        self.sql_ctx = sql_ctx
-        self._sc = sql_ctx._sc
-        self._jschema_rdd = jschema_rdd
-
-        self.is_cached = False
-        self.is_checkpointed = False
-        self.ctx = self.sql_ctx._sc
-        self._jrdd_deserializer = self.ctx.serializer
-
-    @property
-    def _jrdd(self):
-        """
-        Lazy evaluation of PythonRDD object. Only done when a user calls methods defined by the
-        L{RDD} super class (map, count, etc.).
-        """
-        return self.toPython()._jrdd
-
-    @property
-    def _id(self):
-        return self._jrdd.id()
-
-    def saveAsParquetFile(self, path):
-        """
-        Saves the contents of this L{SchemaRDD} as a parquet file, preserving the schema.  Files
-        that are written out using this method can be read back in as a SchemaRDD using the
-        L{SQLContext.parquetFile} method.
-
-        >>> from pyspark.context import SQLContext
-        >>> sqlCtx = SQLContext(sc)
-        >>> rdd = sc.parallelize([{"field1" : 1, "field2" : "row1"},
-        ... {"field1" : 2, "field2": "row2"}, {"field1" : 3, "field2": "row3"}])
-        >>> srdd = sqlCtx.inferSchema(rdd)
-        >>> srdd.saveAsParquetFile("/tmp/test.parquet")
-        >>> srdd2 = sqlCtx.parquetFile("/tmp/test.parquet")
-        >>> srdd2.collect() == srdd.collect()
-        True
-        """
-        self._jschema_rdd.saveAsParquetFile(path)
-
-    def registerAsTable(self, name):
-        """
-        Registers this RDD as a temporary table using the given name.  The lifetime of this temporary
-        table is tied to the L{SQLContext} that was used to create this SchemaRDD.
-
-        >>> from pyspark.context import SQLContext
-        >>> sqlCtx = SQLContext(sc)
-        >>> rdd = sc.parallelize([{"field1" : 1, "field2" : "row1"},
-        ... {"field1" : 2, "field2": "row2"}, {"field1" : 3, "field2": "row3"}])
-        >>> srdd = sqlCtx.inferSchema(rdd)
-        >>> srdd.registerAsTable("test")
-        >>> srdd2 = sqlCtx.sql("select * from test")
-        >>> srdd.collect() == srdd2.collect()
-        True
-        """
-        self._jschema_rdd.registerAsTable(name)
-
-    def toPython(self):
-        jrdd = self._jschema_rdd.javaToPython()
-        # TODO: This is inefficient, we should construct the Python Row object
-        # in Java land in the javaToPython function. May require a custom
-        # pickle serializer in Pyrolite
-        return RDD(jrdd, self._sc, self._sc.serializer).map(lambda d: Row(d))
 
 def _test():
     import doctest
