+++ python/run-tests
@@ -28,6 +28,9 @@ FAILED=0
 
 rm -f unit-tests.log
 
+# Remove the metastore and warehouse directory created by the HiveContext tests in SparkSQL
+rm -rf metastore warehouse
+
 function run_test() {
     SPARK_TESTING=0 $FWDIR/bin/pyspark $1 2>&1 | tee -a > unit-tests.log
     FAILED=$((PIPESTATUS[0]||$FAILED))
@@ -46,6 +49,7 @@ function run_test() {
 run_test "pyspark/rdd.py"
 run_test "pyspark/context.py"
 run_test "pyspark/conf.py"
+run_test "pyspark/sql.py"
 run_test "-m doctest pyspark/broadcast.py"
 run_test "-m doctest pyspark/accumulators.py"
 run_test "-m doctest pyspark/serializers.py"
@@ -55,9 +59,6 @@ run_test "pyspark/mllib/classification.py"
 run_test "pyspark/mllib/clustering.py"
 run_test "pyspark/mllib/recommendation.py"
 run_test "pyspark/mllib/regression.py"
-
-# Remove the metastore directory created by the HiveContext tests in SparkSQL
-rm -r metastore
 
 if [[ $FAILED == 0 ]]; then
     echo -en "\033[32m"  # Green
